// frontend/src/features/vault/views/vault-table/types.ts
export interface TableMetadata extends Readonly<Record<string, unknown>> {
  readonly parent_id?: string;
  readonly table_id?: string;
  readonly database_table_id?: string;
  readonly database_id?: string;
  readonly icon?: string;
  readonly translation_lang?: string;
  readonly translation_stale?: boolean;
}
export interface TableNote extends Record<string, unknown> {
  readonly id: string;
  readonly title?: string;
  readonly parent_id?: string;
  readonly resolved_table_id?: string;
  readonly last_modified?: string;
  readonly metadata?: TableMetadata | null;
}